project/requirements.txt
wikipedia==1.4.0
Wikipedia-API==0.6.0
spotipy==2.23.0
python-dotenv==1.0.0
